feat:protect routes of posts

+++ app/Http/Controllers/PostController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Gate;
 use App\Models\Post;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use App\Http\Exceptions\Handler;
-use App\Http\Requests\UpdatePostRequest;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -30,11 +32,13 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
         try {
-            $validatedData = $request->validated();
-            $post = Post::create($validatedData);
+            Gate::authorize('create', Post::class);
+            $post_data = $request->all();   
+            $post_data['user_id'] = Auth::user()->id;
+            $post = Post::create($post_data);
             return new PostResource($post);
         } catch (\Exception $e) {
-            return $this->handler->render($request,$e);
+            return $this->handler->render($request, $e);
         }
     }
 
@@ -42,41 +46,48 @@ class PostController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Post $posts)
+    public function show(Post $post)
     {
-        return new PostResource($posts);
+        return new PostResource($post);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdatePostRequest $request, Post $post)
+    public function update(Request $request, Post $post)
     {
-        
+
         try {
-        $request_params = $request->all();
-        $post->title =  $request_params['title'] ?? $post->title;
-        $post->description = $request_params['description']?? $post->description;
-        $post->category_id = $request_params['category_id']?? $post->category_id;
-        $post->work_type = $request_params['work_type'] ?? $post->work_type;
-        $post->deadline = $request_params['deadline'] ?? $post->deadline;
-        $post->location = $request_params['location'] ?? $post->location;
-        $post->qualifications = $request_params['qualifications'] ?? $post->qualifications;
-        $post->skills = $request_params['skills'] ?? $post->skills;
-        $post->salary_range = $request_params['salary_range'] ?? $post->salary_range;
-        $post->responsibilities = $request_params['responsibilities'] ?? $post->responsibilities;
-        $post->update();
-        return response()->json(['message'=>'post updated successfully','data'=>$post],200);
-        }catch(\Exception $e){
-            return $this->handler->render($request,$e);
+
+            Gate::authorize('update', $post);
+            $request_params = $request->all();
+            $post->title =  $request_params['title'] ?? $post->title;
+            $post->description = $request_params['description'] ?? $post->description;
+            $post->category_id = $request_params['category_id'] ?? $post->category_id;
+            $post->work_type = $request_params['work_type'] ?? $post->work_type;
+            $post->deadline = $request_params['deadline'] ?? $post->deadline;
+            $post->location = $request_params['location'] ?? $post->location;
+            $post->qualifications = $request_params['qualifications'] ?? $post->qualifications;
+            $post->skills = $request_params['skills'] ?? $post->skills;
+            $post->salary_range = $request_params['salary_range'] ?? $post->salary_range;
+            $post->responsibilities = $request_params['responsibilities'] ?? $post->responsibilities;
+            $post->update();
+            return new PostResource($post);
+        } catch (\Exception $e) {
+            return $this->handler->render($request, $e);
         }
     }
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Post $post)
+    public function destroy(Request $request, Post $post)
     {
+        try {
+        Gate::authorize('delete', $post);
         $post->delete();
         return response()->json(['message' => 'post deleted successfully'], 204);
+        } catch (\Exception $e) {
+            return $this->handler->render($request, $e);
+        }
     }
 }
